Refactored store into different files
