+++ frontend/src/store/actions.js
@@ -1,0 +1,1 @@
+export default {};
