+++ frontend/src/store/mutations.js
@@ -1,0 +1,10 @@
+export default {
+  signIn: function(state, jwt) {
+    state.jwt = jwt;
+    state.logged = true;
+  },
+  signout: function(state) {
+    state.jwt = "";
+    state.logged = false;
+  }
+};
