+++ frontend/src/store/state.js
@@ -1,0 +1,4 @@
+export default {
+  jwt: "",
+  logged: false
+};
